+++ test/robotlegs/bender/utilities/macrobot/impl/signalsMappedToMacro.test.ts
@@ -16,8 +16,13 @@ import { IContext, IInjector, Context, CommandMapper } from "@robotlegsjs/core";
 import { SignalCommandTrigger } from "@robotlegsjs/signalcommandmap";
 
 import { DelaySignal } from "../support/DelaySignal";
+import { NoParametersSignal } from "../support/NoParametersSignal";
 import { ParametersSignal } from "../support/ParametersSignal";
+import { TestNestedSequenceBySignalCommand } from "../support/TestNestedSequenceBySignalCommand";
+import { TestParallelByNoPayloadSignalCommand } from "../support/TestParallelByNoPayloadSignalCommand";
 import { TestParallelBySignalCommand } from "../support/TestParallelBySignalCommand";
+import { TestSequenceByCustomPayloadSignalCommand } from "../support/TestSequenceByCustomPayloadSignalCommand";
+import { TestSequenceByNoPayloadSignalCommand } from "../support/TestSequenceByNoPayloadSignalCommand";
 import { TestSequenceBySignalCommand } from "../support/TestSequenceBySignalCommand";
 
 describe("SignalsMappedToMacro", () => {
@@ -73,6 +78,38 @@ describe("SignalsMappedToMacro", () => {
         assert.deepEqual(reported, expected.concat(expected));
     });
 
+    it("payload_dispatched_by_signal_is_mapped_into_nested_sequence_sub_commands", () => {
+        const expected: any[] = [true, 999, "I'm a string!", Symbol("symbol"), { x: 5, y: 5 }, new Date(), [1, 2, 3, 4, 5, 6, 7, 8, 9]];
+
+        signal = new ParametersSignal();
+        injector.bind(Signal).toConstantValue(signal);
+        mapper = signalCommandTrigger.createMapper();
+        mapper.toCommand(TestNestedSequenceBySignalCommand);
+        signal.dispatch.apply(signal, expected);
+
+        assert.deepEqual(reported, expected.concat(expected));
+    });
+
+    it("no_payload_dispatched_by_signal_is_mapped_into_sequence_sub_commands", () => {
+        signal = new NoParametersSignal();
+        injector.bind(Signal).toConstantValue(signal);
+        mapper = signalCommandTrigger.createMapper();
+        mapper.toCommand(TestSequenceByNoPayloadSignalCommand);
+        signal.dispatch.apply(signal);
+
+        assert.deepEqual(reported, []);
+    });
+
+    it("subcommand_payload_is_mapped_into_nested_subcommands", () => {
+        signal = new NoParametersSignal();
+        injector.bind(Signal).toConstantValue(signal);
+        mapper = signalCommandTrigger.createMapper();
+        mapper.toCommand(TestSequenceByCustomPayloadSignalCommand);
+        signal.dispatch.apply(signal);
+
+        assert.deepEqual(reported, ["command1", "command2"]);
+    });
+
     it("payload_dispatched_by_signal_is_mapped_into_parallel_sub_commands", (done: Function) => {
         signal = new DelaySignal();
         injector.bind(Signal).toConstantValue(signal);
@@ -92,6 +129,19 @@ describe("SignalsMappedToMacro", () => {
                 "Complete execution of Command 4"
             ]);
 
+            done();
+        }, 250);
+    });
+
+    it("no_payload_dispatched_by_signal_is_mapped_into_parallel_sub_commands", (done: Function) => {
+        signal = new NoParametersSignal();
+        injector.bind(Signal).toConstantValue(signal);
+        mapper = signalCommandTrigger.createMapper();
+        mapper.toCommand(TestParallelByNoPayloadSignalCommand);
+        signal.dispatch();
+
+        setTimeout(() => {
+            assert.deepEqual(reported, []);
             done();
         }, 250);
     });