+++ test/robotlegs/bender/utilities/macrobot/support/TestNestedSequenceBySignalCommand.ts
@@ -9,12 +9,12 @@ import { injectable } from "@robotlegsjs/core";
 
 import { SequenceMacro } from "../../../../../../src/robotlegs/bender/utilities/macrobot/impl/SequenceMacro";
 
-import { CallbackParametersCommand } from "./CallbackParametersCommand";
+import { CallbackParametersSequenceCommand } from "./CallbackParametersSequenceCommand";
 
 @injectable()
-export class TestSequenceBySignalCommand extends SequenceMacro {
+export class TestNestedSequenceBySignalCommand extends SequenceMacro {
     public prepare(): void {
-        this.add(CallbackParametersCommand);
-        this.add(CallbackParametersCommand);
+        this.add(CallbackParametersSequenceCommand);
+        this.add(CallbackParametersSequenceCommand);
     }
 }